% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsplines.R
\name{bspline}
\alias{bspline}
\title{Create a b-spline design matrix.}
\usage{
bspline(x, xl, xr, ndx, bdeg)
}
\arguments{
\item{x}{The points at which the b-spline is evaluated.}

\item{xl}{The lower bound of the domain.}

\item{xr}{The upper bound of the domain.}

\item{ndx}{The granularity of the basis.}

\item{bdeg}{The degree of the B-spline polynomial components.}
}
\description{
Create a b-spline design matrix.
}
\author{
pschulam
}

